ptrlist changed collected in separate file.   The file is executed after existing files in that folder as RangeVarPlot test seems to interfere with some "python -m pytest ." tests.
